feat(string-split): exportToSqlite3 supports int{32, 64}, but not int

+++ src/diesl/extensions/sqlite.nim
@@ -21,7 +21,7 @@ proc extractOne(input: string, regex: string): string {.exportToSqlite3.} =
 proc rReplace(input: string, old: string, nw: string): string {.exportToSqlite3.} = 
   return re.replace(input, re(old), nw)
 
-proc stringSplit(input: string, splitOn: string, index: int): string {.exportToSqlite3.} = 
+proc stringSplit(input: string, splitOn: string, index: int64): string {.exportToSqlite3.} = 
   let ss = input.split(input, re(splitOn))
   return if index < len(ss):
     ss[index]
